Remove let and Assignments
