+++ docs/reference.md
@@ -9,7 +9,6 @@ Types
 ::: funml.types
     options:
         members:
-            - Assignment
             - Expression
             - MatchExpression
             - MLType
